cost_analyzer/requirements.txt
google-cloud-bigquery
google-cloud-monitoring
google-api-python-client
requests

